requirements.txt
flask
python-dotenv
llama-parse
langchain-openai
faiss-cpu
langchain
werkzeug
